backend/requirements.txt
Flask
sqlalchemy
marshmallow
flask_cors
PyJWTmarshmallow
